@@ -1,25 +1,29 @@
-module github.com/boreq/bolt-ui
+module github.com/tnngo/bolt-ui
+
+go 1.21.9
 
 require (
-	github.com/NYTimes/gziphandler v1.1.0
+	github.com/NYTimes/gziphandler v1.1.1
 	github.com/boreq/errors v0.1.0
-	github.com/boreq/guinea v0.0.0-20190218203212-75c10cec45e9
+	github.com/boreq/guinea v0.1.0
 	github.com/boreq/rest v0.1.0
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-stack/stack v1.8.0 // indirect
-	github.com/google/wire v0.5.0
-	github.com/inconshreveable/log15 v0.0.0-20180818164646-67afb5ed74ec
-	github.com/julienschmidt/httprouter v1.2.0
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
-	github.com/oklog/ulid/v2 v2.0.2
-	github.com/pkg/errors v0.8.1
-	github.com/rs/cors v1.6.0
-	github.com/stretchr/testify v1.7.0
-	go.etcd.io/bbolt v1.3.3
-	golang.org/x/sys v0.0.0-20210806184541-e5e7981a1069 // indirect
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	github.com/google/wire v0.6.0
+	github.com/inconshreveable/log15 v2.16.0+incompatible
+	github.com/julienschmidt/httprouter v1.3.0
+	github.com/oklog/ulid/v2 v2.1.0
+	github.com/pkg/errors v0.9.1
+	github.com/rs/cors v1.11.0
+	github.com/stretchr/testify v1.9.0
+	go.etcd.io/bbolt v1.3.9
 )
 
-go 1.16
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/mattn/go-colorable v0.1.2 // indirect
+	github.com/mattn/go-isatty v0.0.8 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/term v0.16.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
